+++ example/example.dart
@@ -1,7 +1,7 @@
-import 'package:dart_pgmq/src/pgmq/database_connection.dart';
-import 'package:dart_pgmq/src/pgmq/pgmq.dart';
+import 'package:dart_pgmq/dart_pgmq.dart';
 
-void main() async {
+Future<void> main() async {
+  // Create a DatabaseConnection
   final databaseParam = DatabaseConnection(
       host: 'localhost',
       database: 'postgres',
@@ -9,49 +9,26 @@ void main() async {
       username: 'postgres',
       ssl: false,
       port: 5460);
-  print("Start ...");
-  try {
-    final pgmq = await Pgmq.createConnection(param: databaseParam);
 
-    final queue = await pgmq.createQueue(queueName: 'yaya');
-    // Purge queue
-    await queue.purgeQueue();
+  // Create a connexion
+  final pgmq = await Pgmq.createConnection(param: databaseParam);
 
-    final (pausableTimer1, stream1) =
-        queue.pausablePull(duration: Duration(milliseconds: 100));
-    pausableTimer1.start();
+  //  Create a queue
+  final queue = await pgmq.createQueue(queueName: 'queueName');
 
-    final (pausableTimer2, stream2) =
-        queue.pausablePull(duration: Duration(milliseconds: 100));
-    pausableTimer2.start();
-
-    stream1.listen((event) async {
-      print("Subscription 1: ${event.messageID}");
-      print("Suscription 1 Deleted : ${await queue.delete(event.messageID)}");
-    });
-
-    stream2.listen((event) async {
-      print("Subscription 2: ${event.messageID}");
-      print("Suscription 2 Deleted : ${await queue.delete(event.messageID)}");
-    });
-
-    Future.delayed(Duration(seconds: 1), pausableTimer1.pause);
-
-    Future.delayed(Duration(seconds: 10), pausableTimer1.start);
-
-    for (var i = 1; i <= 20; i++) {
-      await Future.delayed(Duration(seconds: 3));
-      final payload = {'id': i, 'message': 'message $i'};
-      await queue.send(payload);
-    }
-
-    // final data = (await queue.read(maxReadNumber: 5));
-
-    // for (final msg in data ?? <Message>[]) {
-    //   print(msg.payload);
-    // }
-  } catch (e, stackTrace) {
-    print(stackTrace);
-    print(e.toString());
+  // Send message
+  for (var i = 1; i <= 20; i++) {
+    await Future.delayed(Duration(seconds: 3));
+    final payload = {'id': i, 'message': 'message $i'};
+    await queue.send(payload);
   }
+
+  // Read message
+  final data = (await queue.read(maxReadNumber: 5));
+  for (final msg in data ?? <Message>[]) {
+    print(msg.payload);
+  }
+
+  // Purge queue
+  await queue.purgeQueue();
 }