+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
-- Initial version.
+Initial version.
+
+- feat: Includes the ability to create a `PgmqConnexion`.
+- feat: Includes the ability to create, purge, drop a `Queue`.
+- feat: Includes the ability to send, read, pop, archive, delete a `Message`.
+- feat: Includes the ability to set a `vtc`.